+++ Templates/Swift/Sources/Client/APIClient.swift
@@ -3,6 +3,12 @@
 import Foundation
 import Alamofire
 
+#if SPM_SPLIT_MODE_ON
+import {{ options.name }}Requests
+import {{ options.name }}Models
+import {{ options.name }}SharedCode
+#endif
+
 /// Manages and sends APIRequests
 public class APIClient {
 
@@ -175,7 +181,7 @@ public class APIClient {
             result = .failure(apiError)
             requestBehaviour.onFailure(error: apiError)
         }
-        let response = APIResponse<T>(request: request, result: result, urlRequest: dataResponse.request, urlResponse: dataResponse.response, data: dataResponse.data, timeline: dataResponse.timeline)
+        let response = APIResponse<T>(request: request, result: result, urlRequest: dataResponse.request, urlResponse: dataResponse.response, data: dataResponse.data, timeline: Timeline(requestStartTime: dataResponse.timeline.requestStartTime, initialResponseTime: dataResponse.timeline.initialResponseTime, requestCompletedTime: dataResponse.timeline.requestCompletedTime, serializationCompletedTime: dataResponse.timeline.serializationCompletedTime))
         requestBehaviour.onResponse(response: response.asAny())
 
         completionQueue.async {
@@ -207,43 +213,5 @@ extension APIRequest {
     /// makes a request using the default APIClient. Change your baseURL in APIClient.default.baseURL
     public func makeRequest(complete: @escaping (APIResponse<ResponseType>) -> Void) {
         APIClient.default.makeRequest(self, complete: complete)
-    }
-}
-
-// Create URLRequest
-extension APIRequest {
-
-    /// pass in an optional baseURL, otherwise URLRequest.url will be relative
-    public func createURLRequest(baseURL: String = "", encoder: RequestEncoder = JSONEncoder()) throws -> URLRequest {
-        let url = URL(string: "\(baseURL)\(path)")!
-        var urlRequest = URLRequest(url: url)
-        urlRequest.httpMethod = service.method
-        urlRequest.allHTTPHeaderFields = headers
-
-        // filter out parameters with empty string value
-        var queryParams: [String: Any] = [:]
-        for (key, value) in queryParameters {
-            if String.init(describing: value) != "" {
-                queryParams[key] = value
-            }
-        }
-        if !queryParams.isEmpty {
-            urlRequest = try URLEncoding.queryString.encode(urlRequest, with: queryParams)
-        }
-
-        var formParams: [String: Any] = [:]
-        for (key, value) in formParameters {
-            if String.init(describing: value) != "" {
-                formParams[key] = value
-            }
-        }
-        if !formParams.isEmpty {
-            urlRequest = try URLEncoding.httpBody.encode(urlRequest, with: formParams)
-        }
-        if let encodeBody = encodeBody {
-            urlRequest.httpBody = try encodeBody(encoder)
-            urlRequest.setValue("application/json", forHTTPHeaderField: "Content-Type")
-        }
-        return urlRequest
     }
 }