+++ Templates/Swift/Sources/Requests/APIResponse.swift
@@ -1,7 +1,15 @@
 {% include "Includes/Header.stencil" %}
 
 import Foundation
-import Alamofire
+
+#if canImport(FoundationNetworking)
+import FoundationNetworking
+#endif
+
+#if SPM_SPLIT_MODE_ON
+import {{ options.name }}Models
+import {{ options.name }}SharedCode
+#endif
 
 public protocol APIResponseValue: CustomDebugStringConvertible, CustomStringConvertible {
     associatedtype SuccessType{% if options.codableResponses %} : Codable{% endif %}
@@ -59,7 +67,7 @@ public struct APIResponse<T: APIResponseValue> {
     /// The timeline of the complete lifecycle of the request.
     public let timeline: Timeline?
 
-    init(request: APIRequest<T>, result: APIResult<T>, urlRequest: URLRequest? = nil, urlResponse: HTTPURLResponse? = nil, data: Data? = nil, timeline: Timeline? = nil) {
+    public init(request: APIRequest<T>, result: APIResult<T>, urlRequest: URLRequest? = nil, urlResponse: HTTPURLResponse? = nil, data: Data? = nil, timeline: Timeline? = nil) {
         self.request = request
         self.result = result
         self.urlRequest = urlRequest