+++ Sources/Swagger/Schema/Discriminator.swift
@@ -5,6 +5,11 @@ public struct Discriminator {
 
     public var propertyName: String
     public let mapping: [String: String]
+
+    public init(propertyName: String, mapping: [String: String]) {
+        self.propertyName = propertyName
+        self.mapping = mapping
+    }
 }
 
 extension Discriminator: JSONObjectConvertible {