+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 - Fixed inline allOf group generation
 - Fixed property generation when there is only one group schema;  the first group schema type will be used as the type #217
+- Added `anyType` option that allows to override `Any` in models
 
 ## 4.3.1