+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next Version
 
+### Added
+- MediaItem schema is now defined as a possible reference in order to generate models of each server response. #239 @alexruperez
+
 ### Fixed
 - Fixed inline allOf group generation
 - Fixed property generation when there is only one group schema;  the first group schema type will be used as the type #217