+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fixed property generation when there is only one group schema;  the first group schema type will be used as the type #217
 - Added `anyType` option that allows to override `Any` in models
 - Fixed date encoding formatter to conform to RFC3339
+- Fixed `.swift-version` to use Swift 5.2 instead of Swift 4.1 #246
 
 ## 4.3.1