+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 - Support multiple authentication types per operation #222 @malteburkert
+- Added support for `decimal` number format #259 @yonaskolb
+- Added `numberType`, `doubleType`, `floatType` and `decimalType` template options #259 @yonaskolb
 
 ### Fixed
 - Fixed inline allOf group generation